Small commit. Only exists 'cause Vercel is strange
Basically this commmit doesnt't add much, just exists 'cause Vercel for
some reason forgets (I think) the routing of the pages and when you
refresh any of them the whole thing breaks and you gotta go to the main
page and reload there to get it fixed. Really strange, so strange in
fact that not even ChatGPT was really sure of what do. Aside from this,
I just made the overlay function on the posts be its own file.