Actual commit with important stuff
This commit mainly aims to centralizing the posting system. Now I have
to do before making a new post is just adding 7 lines and changing the
post location. I plan on also making it even better by idk identifying
the function it is currently running and outputing that in the file
location but that is for another day. Right now my main fear is that
Vercel will stop identifying the posts' location again for no reason.